All components separated and renamed svgs
